Updated mag character and added animations

+++ Assets/Scripts/Skills/SkillsParameters.cs
@@ -2,5 +2,5 @@
 public class SkillsParameters
 {
     public float Countdown;
-    public float DamageMin, DamageMax;
+    public int DamageMin, DamageMax;
 }
